requirements.txt
openai
wolframalpha
flask
path.py